re-group the camera controls and get rid of extra code. make the layout more similar to other sections
