+++ agave_app/CMakeLists.txt
@@ -15,8 +15,6 @@ target_sources(agaveapp PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/agaveGui.ui"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Aperture.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Aperture.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/ApertureWidget.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/ApertureWidget.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceDockWidget.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceDockWidget.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppearanceSettingsWidget.cpp"
@@ -35,12 +33,8 @@ target_sources(agaveapp PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/Controls.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Film.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Film.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/FilmWidget.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/FilmWidget.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Focus.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Focus.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/FocusWidget.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/FocusWidget.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLView3D.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/GLView3D.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/loadDialog.cpp"
@@ -49,8 +43,6 @@ target_sources(agaveapp PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/mainwindow.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Projection.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Projection.h"
-	"${CMAKE_CURRENT_SOURCE_DIR}/ProjectionWidget.cpp"
-	"${CMAKE_CURRENT_SOURCE_DIR}/ProjectionWidget.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/QRenderSettings.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/QRenderSettings.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RangeWidget.cpp"
